@@ -641,7 +641,7 @@
         "shelljs": "^0.8.5",
         "source-map-support": "^0.5.21",
         "ts-node": "^10.9.2",
-        "typescript": "^5.0.3",
+        "typescript": "^5.8.2",
         "vsce": "^1.87.0",
         "webpack": "^5.74.0",
         "webpack-cli": "^5.1.4"
